pyproject.toml
[tool.poetry]
name = "organizer"
version = "0.1.0"
description = ""
authors = ["myuR <myurpvn@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
structlog = "^24.4.0"
typer = "^0.12.5"
pyyaml = "^6.0.2"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
app = "organizer.__main__:app"